refactor(workbench): extract composer column into WorkbenchComposerColumn
Move the permission bar and composer input shell out of AppWorkbench so composer layout can change without opening the 18k-line host. Send/queue verbs stay on the host.
